+++ lib/dialogs.dart
@@ -6,4 +6,5 @@ library dialogs;
 export 'src/widgets/dialogs/about_app_dialog.dart';
 export 'src/widgets/dialogs/lit_titled_dialog.dart';
 export 'src/widgets/dialogs/lit_date_picker_dialog.dart';
+export 'src/widgets/dialogs/lit_color_picker_dialog.dart';
 export 'src/widgets/dialogs/discard_draft_dialog.dart';